requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
cfgv==3.3.0
decorator==5.0.9
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-mdeditor==0.1.20
django-seed==0.2.2
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
dj_database_url
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.13
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
nodeenv==1.6.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
pre-commit==2.14.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
virtualenv==20.7.2
wcwidth==0.2.5
